% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{make a summary table of dataset}
\usage{
summaryTable(data_tips)
}
\arguments{
\item{data_tips}{A dataframe that has been cleaned}
}
\value{
A tibble with all kinds of information
}
\description{
make a summary table of dataset
}
\examples{
url <-  paste0(
"https://raw.githubusercontent.com/JustinTrenchcoat/",
"STAT_301_Personal_Project/refs/heads/main/tips.csv"
)
data_tips <- read.csv(url(url), header=TRUE)
summary <- summaryTable(data_tips)

}
